faqs_project/requirements.txt
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.5
django-ckeditor==6.7.2
django-heroku==0.3.1
django-js-asset==3.0.1
django-redis==5.4.0
djangorestframework==3.15.2
docker==7.1.0
flake8==7.1.1
googletrans==4.0.2
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.4
pywin32==308
redis==5.2.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.8.2
